pyproject.toml
[project]
name = "aso_agent"
version = "0.0.1"
description = "ASO Analysis Agent with Web Service Interface"
authors = [
    { name = "William Fu-Hinthorn", email = "13333726+hinthornw@users.noreply.github.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    # Core LangGraph
    "langgraph>=0.2.6",
    "langgraph-checkpoint-sqlite>=2.0.1",
    "langchain-core>=0.3.33",
    "langchain-openai>=0.3.27",
    
    # Web Service
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "streamlit>=1.40.1",
    "httpx>=0.27.2",
    
    # Data & Async
    "pydantic>=2.0.0",
    "pydantic-settings>=2.6.1",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.20.0",
    "pandas>=2.2.0",
    
    # Visualization
    "plotly>=5.17.0",
    
    # Environment
    "python-dotenv>=1.0.1",
    "nest-asyncio>=1.6.0",
    
    # Existing ASO tools
    "browser-use>=0.5.4",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent", "lib", "service", "client", "schema", "memory"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"
"lib" = "src/lib"
"service" = "src/service"
"client" = "src/client"
"schema" = "src/schema"
"memory" = "src/memory"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.2.8",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
